libsystem: Added support for 64bit numbers.

+++ libraries/libsystem/utils/NumberFormatter.cpp
@@ -5,7 +5,7 @@
 static const char *digits = "0123456789abcdefghijklmnopqrstuvwxyz";
 static const char *digits_capitalized = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
-size_t format_uint(NumberFormater formater, unsigned int value, char *str, size_t size)
+size_t format_uint(NumberFormater formater, unsigned long value, char *str, size_t size)
 {
     if (size == 0)
         return 0;
@@ -48,7 +48,7 @@ size_t format_uint(NumberFormater formater, unsigned int value, char *str, size_
     return written;
 }
 
-size_t format_int(NumberFormater formater, int value, char *str, size_t size)
+size_t format_int(NumberFormater formater, long value, char *str, size_t size)
 {
     if (size == 0)
         return 0;
